Fix linter errors, warning
