+++ package.json
@@ -18,10 +18,15 @@
     "react-datepicker": "^4.2.0",
     "react-dom": "^17.0.2",
     "react-dropdown-select": "^4.7.4",
+    "react-icons": "^4.2.0",
     "react-redux": "^7.2.4",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-styled-select": "^3.0.0",
     "redux": "^4.1.1",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.3.0",
     "styled-components": "^4.4.1",
     "web-vitals": "^1.0.1"
   },
