+++ src/App.js
@@ -1,10 +1,10 @@
-/* eslint-disable no-undef */
 import React, { useEffect } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import logIn from './redux/actions/userAction';
 import SignInPage from './pages/sign-in-page/SignInPage';
 import HomePage from './pages/home-page/HomePage';
+import { requestFacilityData } from './axios/requests';
 import AppointmentsPage from './pages/appointments-page/AppointmentPage';
 import Navbar from './components/navbar/Navbar';
 import LogInPage from './pages/log-in-page/LogInPage';
