+++ src/axios/requests.js
@@ -2,7 +2,8 @@ import Axios from 'axios';
 import {
   fetchFacilitySuccess,
   fetchFacilityFailure,
-} from '../redux/actions/facilityAction';
+} from '../redux/actions/FacilityAction';
+
 import logIn from '../redux/actions/userAction';
 import fetchAppointments from '../redux/actions/AppointAction';
