+++ src/components/appointments-table/AppointmentsTable.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import { requestAppointments } from '../../axios/requests';
 import styles from './styles/appointments.module.css';
 
 const AppointmentsTable = () => {
