+++ src/components/dropdown/styles/dropdown.js
@@ -1,0 +1,12 @@
+import Select from 'react-dropdown-select';
+import styled from 'styled-components/macro';
+
+const StyledSelect = styled(Select)`
+ color: white !important;
+    border: 1px solid white !important;
+    border-radius: 20px !important;
+    max-width: 180px;
+    min-width: 150px
+`;
+
+export default StyledSelect;
