+++ src/components/log-out/LogOut.js
@@ -1,3 +1,5 @@
+/* eslint-disable  jsx-a11y/click-events-have-key-events */
+/* eslint-disable  jsx-a11y/no-noninteractive-element-interactions */
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { NavLink } from 'react-router-dom';
