+++ src/pages/sign-in-page/SignInPage.js
@@ -1,7 +1,8 @@
+/* eslint-disable  react/no-array-index-key */
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Redirect } from 'react-router';
-import styles from './styles/signInPage.module.css';
+import styles from './styles/SignInPage.module.css';
 import InputField from '../../components/input-field/InputField';
 import Button from '../../components/button/Button';
 import { requestSignUserIn } from '../../axios/requests';
